changes to minikore
