+++ src/main/scala/org/kframework/minikore/MiniToText.scala
@@ -1,7 +1,7 @@
 package org.kframework.minikore
 
 import org.apache.commons.lang3.StringEscapeUtils
-import org.kframework.minikore.NewKore._
+import org.kframework.minikore.MiniKore._
 
 /** Function (i.e., unparser) from MiniKore to String. */
 object MiniToText {
